+++ formio/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 16.0.3.1
+
+Improve app/module installation, by downloading and installing a default formio.js library version after install.\
+Also log a warning in case an error (eg ConnectionError) occurs.
+
 ## 16.0.3.0
 
 1. Refactored the Form its Form Builder (field) domain/filter.